+++ CHANGELOG.md
@@ -8,6 +8,7 @@ A major update to the BOOMER algorithm that introduces the following changes:
 * The (label- and example-wise) logistic loss functions have been rewritten to better prevent numerical problems.
 * Approximate methods for evaluating potential conditions of rules, based on unsupervised binning methods (currently equal-width- and equal-frequency-binning), have been added.
 * The algorithm does now support data sets with missing feature values.
+* The loss function `label-wise-squared-hinge-loss` has been added. 
 * Experiments using single-label data sets are now supported out of the box.
 
 ### Version 0.3.0 (Sep 14, 2020)